third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/named-access-on-the-window-object/cross-origin-named-access.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Accessing a cross-origin child global by name returns undefined
  assert_equals: window.frame1 should be undefined for cross-origin iframes expected (undefined) undefined but got (object) [stringifying object threw SecurityError: Failed to read a named property from 'Window': Blocked a frame with origin "http://web-platform.test:8001" from accessing a cross-origin frame. with type object]
[FAIL] Child accessing its own global through a cross-origin parent's WindowProxy throws an exception
  assert_equals: Accessing `window.parent.frame1` from the child throws a SecurityError expected "SecurityError" but got "No exception"
Harness: the test ran to completion.
